Feature / Discount tag checkout view (#705)
* Move DiscountTag from OfferNew to own component

* Add DiscountTag to checkout and add margins in both Checkout and Sidebar

* Move isNorwegianBundle to utils

* Update src/client/pages/OfferNew/utils.ts
