+++ src/client/pages/OfferNew/utils.ts
@@ -298,6 +298,9 @@ export const isNoDiscount = (campaigns: Campaign[]) =>
     campaigns[0].incentive.__typename === 'NoDiscount') ||
   false
 
+export const isNorwegianBundle = (offerData: OfferData): boolean =>
+  isBundle(offerData) && isNorwegian(offerData)
+
 // TODO: Make Contract Types more generic
 export type HomeInsuranceTypeOfContract = Exclude<
   TypeOfContract,
